similar attribute search regardless of the parent
Change the default behaviour of "getAttributes" to always include the
color attribute.
"findSimilarAttr", now takes an additional parameter as a list of column
names for the creation of sets, this makes it more extendable for future
additions.
Add function description to getAttributes and findSimilarAttr.
Add a additional Exception case to "getProperties" to check for a
KeyError.