Bug Fix: product-import.py Empty TK root window
Fix a bug, where an empty Tkinter window was present at the start of
the script. Use the withdraw method on the root window to hide it, when
askopenfile is called.
Destroy the tkinter main window at the end of the script.

Additionally, skip wait for input for non windows systems.