Remove manual upload of extern id & export files
Adjust the config to take two additional fields for a extern ID file and
the URL of the plentymarkets Export required for the Variation ID.
Refactor the get_variation_id function to work with the pandas library

Create function get_externalid, which retrieves the external id stored
within a local file for each sku.

Add config handling in product_import for the new fields.