item_upload: various adjustments
Remove the old way of handling package measurements, which involved
finding a set of complete values within an article and using it for all
variations. Use the individual values and use empty values if required,
additionally remove the conversion to milimeter.

The marking field was removed so remove it here as well and at the name
3 field.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>