Rework Item upload function
Use pandas instead of the csv module.
Clean up the function to be less clunky and less repetitive.
Improve the performance of the function by using the pandas module
instead of iterating multiple times over the same data set.
Make the code easier to maintain and read by adding constants which
represent the column names to be used, when one of those column names
should change it has to be changed at a single location.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>