Bug Fix: item_upload.py: Windows PermissionError
When item_upload writes into a file that is used by
another process windows systems give PermissionErrors.
To accommodate for this, we instead wrap the write function
into a while loop, where we check if the write function gives
a PermissionError, if so we give the user an error message so
he can fix it manually and then redo the loop.

Signed-off-by: Jan Sallermann <jan.sallermann@panasiam.de>
Reviewed-by: Sebastian Fricke <sebastian.fricke@posteo.net>